# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MBKM API” Manajemen User & Role",
    "version": "1.1.0",
    "description": "Admin mengelola akun dosen (Tim MBKM, GPM, Penguji Semhas, DPL, dll - satu dosen boleh punya beberapa role) dan akun mahasiswa langsung dibuat oleh admin. Klik Authorize lalu isi token dari /auth/login untuk mencoba endpoint yang butuh login."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "meta": {
            "type": "object"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "example": "GantiPasswordIni123"
          }
        }
      },
      "UpdateSelfRequest": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string"
          },
          "alamat": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "CreateCplRequest": {
        "type": "object",
        "required": [
          "kode",
          "program_studi_id",
          "deskripsi"
        ],
        "properties": {
          "kode": {
            "type": "string",
            "example": "CPL-1"
          },
          "program_studi_id": {
            "type": "integer",
            "example": 3
          },
          "deskripsi": {
            "type": "string"
          }
        }
      },
      "UpdateCplRequest": {
        "type": "object",
        "description": "Minimal satu field",
        "properties": {
          "kode": {
            "type": "string"
          },
          "program_studi_id": {
            "type": "integer"
          },
          "deskripsi": {
            "type": "string"
          }
        }
      },
...
      "CreateMataKuliahRequest": {
        "type": "object",
        "required": [
          "program_studi_id",
          "kode",
          "nama",
          "cpl_id",
          "sks"
        ],
        "properties": {
          "program_studi_id": {
            "type": "integer",
            "example": 3
          },
          "kode": {
            "type": "string",
            "example": "KFU"
          },
          "nama": {
            "type": "string",
            "example": "Algoritma Evolusi (KFU1202)"
          },
          "cpl_id": {
            "type": "string",
            "format": "uuid",
            "description": "Harus CPL milik program studi yang sama"
          },
          "sks": {
            "type": "integer",
            "minimum": 1,
            "maximum": 6,
            "example": 3
          },
          "cpmk": {
            "type": "array",
            "description": "Deskripsi CPMK, jumlah bebas (maks 50). Isian kosong diabaikan.",
            "items": {
              "type": "string"
            },
            "example": [
              "CPMK-1: Mahasiswa mampu menjelaskan konsep dasar algoritma evolusi."
            ]
          }
        }
      },
      "UpdateMataKuliahRequest": {
        "type": "object",
        "description": "Minimal satu field. Kalau program_studi_id diganti, kirim juga cpl_id baru. Kalau cpmk dikirim, seluruh CPMK diganti; kalau tidak dikirim, tidak berubah.",
        "properties": {
          "program_studi_id": {
            "type": "integer"
          },
          "kode": {
            "type": "string"
          },
          "nama": {
            "type": "string"
          },
          "cpl_id": {
            "type": "string",
            "format": "uuid"
          },
          "sks": {
            "type": "integer"
          },
          "cpmk": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Berhasil, dapat token JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                }
              }
            }
          },
          "401": {
            "description": "Email/password salah"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Profil & role akun yang sedang login",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Auth"
        ],
        "summary": "Edit profil sendiri (nama/alamat/password)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSelfRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/cpl": {
      "get": {
        "tags": [
          "CPL"
        ],
        "summary": "Daftar CPL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "program_studi_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter program studi. Dipakai dropdown CPL di form mata kuliah."
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK: {rows,total}"
          }
        }
      },
      "post": {
        "tags": [
          "CPL"
        ],
        "summary": "Tambah CPL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCplRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "201": {
            "description": "Dibuat"
          },
          "409": {
            "description": "Kode sudah terdaftar di program studi tsb"
          }
        }
      }
    },
    "/cpl/{id}": {
      "get": {
        "tags": [
          "CPL"
        ],
        "summary": "Detail CPL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID CPL"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Tidak ditemukan"
          }
        }
      },
      "put": {
        "tags": [
          "CPL"
        ],
        "summary": "Edit CPL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID CPL"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCplRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Tidak ditemukan"
          },
          "409": {
            "description": "Kode duplikat"
          },
          "422": {
            "description": "Program studi tidak bisa diubah karena CPL sudah dipakai mata kuliah"
          }
        }
      },
      "delete": {
        "tags": [
          "CPL"
        ],
        "summary": "Hapus CPL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID CPL"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "Dihapus"
          },
          "404": {
            "description": "Tidak ditemukan"
          },
          "409": {
            "description": "CPL masih dipakai mata kuliah"
          }
        }
      }
    },
    "/mata-kuliah": {
      "get": {
        "tags": [
          "Mata Kuliah"
        ],
        "summary": "Daftar mata kuliah (beserta CPL & CPMK)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "program_studi_id",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filter program studi. Dipakai dropdown CPL di form mata kuliah."
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK: {rows,total}"
          }
        }
      },
      "post": {
        "tags": [
          "Mata Kuliah"
        ],
        "summary": "Tambah mata kuliah",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMataKuliahRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "201": {
            "description": "Dibuat"
          },
          "409": {
            "description": "Kode sudah terdaftar di program studi tsb"
          }
        }
      }
    },
    "/mata-kuliah/{id}": {
      "get": {
        "tags": [
          "Mata Kuliah"
        ],
        "summary": "Detail mata kuliah",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID mata kuliah"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Tidak ditemukan"
          }
        }
      },
      "put": {
        "tags": [
          "Mata Kuliah"
        ],
        "summary": "Edit mata kuliah",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID mata kuliah"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMataKuliahRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Tidak ditemukan"
          },
          "409": {
            "description": "Kode duplikat"
          }
        }
      },
      "delete": {
        "tags": [
          "Mata Kuliah"
        ],
        "summary": "Hapus mata kuliah (CPMK ikut terhapus)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID mata kuliah"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin"
          },
          "200": {
            "description": "Dihapus"
          },
          "404": {
            "description": "Tidak ditemukan"
          }
        }
      }
    },
...
  }
}